requirements.txt
aiogram==2.23.1
psycopg2-binary
sqlalchemy[asyncio]
asyncpg